backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pymysql
python-dotenv
mysqlclient
cryptography